Server/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pandas==2.1.1
numpy==1.26.0
requests==2.31.0
google-generativeai==0.3.2
scikit-learn==1.4.1.post1  # Updated to match the pickle file version
gunicorn==21.2.0           # Added for production deployment